Repository y mappers

+++ app/src/main/java/com/example/marketlens/data/results/CoinGeckoResult.kt
@@ -8,7 +8,7 @@ data class CoinGeckoResult(
     @SerializedName("name") val name: String?,
     @SerializedName("image") val image: String?,
     @SerializedName("current_price") val currentPrice: Double?,
-    @SerializedName("market_cap") val marketCap: Long?,
+    @SerializedName("market_cap") val marketCap: Double?,
     @SerializedName("market_cap_rank") val marketCapRank: Int?,
     @SerializedName("price_change_24h") val priceChange24h: Double?,
     @SerializedName("price_change_percentage_24h") val priceChangePercentage24h: Double?,
